fix: size field bytes length is 20

+++ migrations/20220623125000-create-send-links-tables.js
@@ -74,7 +74,7 @@ module.exports = {
         type: Sequelize.STRING(400),
       },
       size: {
-        type: Sequelize.INTEGER.UNSIGNED,
+        type: Sequelize.INTEGER(20).UNSIGNED,
       },
       created_at: {
         type: Sequelize.DATE,
